memperbarui struktur menggunakan MVC
