+++ src/main/java/id/banksampah/Core/Service.java
@@ -1,0 +1,18 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package saldo.core;
+import java.sql.Connection;
+/**
+ *
+ * @author Asus
+ */
+public class Service {
+    protected Connection connection;
+
+    public Service() {
+        this.connection = (new MySQL()).getConnection();
+    }
+}
